requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-allauth==0.63.3
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.2
django-jazzmin==3.0.0
django-modeltranslation==0.19.5
django-redis==5.4.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
easyocr==1.7.1
et-xmlfile==1.1.0
filelock==3.15.4
fsspec==2024.6.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
imageio==2.35.0
inflection==0.5.1
Jinja2==3.1.4
kombu==5.3.7
lazy_loader==0.4
loguru==0.7.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyclipper==1.3.0.post5
pycparser==2.22
PyJWT==2.8.0
pytesseract==0.3.10
python-bidi==0.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.5
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
sympy==1.13.2
tifffile==2024.8.10
torch==2.4.0
torchvision==0.19.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.0.3
whitenoise